requirements.txt
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3.1
Django==2.0.8
django-rest-auth==0.9.3
djangorestframework==3.8.2
dlib==19.10.0
ez-setup==0.9
face-recognition==1.2.2
face-recognition-models==0.3.0
idna==2.6
mysql-connector==2.1.6
numpy==1.14.2
opencv-python==3.4.0.12
Pillow==5.0.0
pycparser==2.18
pygame==1.9.3
PyMySQL==0.9.2
pyTelegramBotAPI==3.6.2
pytz==2018.5
requests==2.18.4
scipy==1.0.1
six==1.11.0
urllib3==1.22
